refact admin folder structure
M  src/app/doctoralCenter/admin/_components/DoctoralCenterAdminEventManagementGrid.jsx
M  src/app/doctoralCenter/admin/_components/DoctoralCenterAdminHomeGrid.jsx
A  src/app/doctoralCenter/admin/_constants/eventConstants.jsx
A  src/app/doctoralCenter/admin/_constants/unauthorizedUsersColumns.jsx
A  src/app/doctoralCenter/admin/_constants/userManagementColumns.jsx
M  src/app/doctoralCenter/admin/_lib/EventManagementGridData.jsx
M  src/app/doctoralCenter/admin/_lib/UnauthorizedUsersGridData.jsx
M  src/app/doctoralCenter/admin/_lib/UserManagementGridData.jsx
M  src/components/main-layout/common/Header.jsx
