+++ src/app/doctoralCenter/admin/_lib/EventManagementGridData.jsx
@@ -37,67 +37,8 @@ export default function EventManagementGridData() {
     }, process.env.NEXT_PUBLIC_FETCH_API_DURATION);
   }, [setRows]);
 
-  const columns = [
-    { field: "description", headerName: "Описание", flex: 1.5, minWidth: 200 },
-    {
-      field: "formattedTimestamp",
-      headerName: "Време",
-      flex: 1,
-      minWidth: 250
-    },
-    {
-      field: "action",
-      headerName: "Действие",
-      headerAlign: "right",
-      align: "right",
-      flex: 1,
-      minWidth: 250
-    },
-    {
-      field: "level",
-      headerName: "Ниво",
-      headerAlign: "right",
-      align: "right",
-      flex: 2,
-      minWidth: 100
-    },
-    {
-      field: "oid",
-      headerName: "Oid",
-      headerAlign: "right",
-      align: "right",
-      flex: 2,
-      minWidth: 300
-    },
-    {
-      field: "name",
-      headerName: "Име",
-      headerAlign: "right",
-      align: "right",
-      flex: 2,
-      minWidth: 300
-    },
-    {
-      field: "email",
-      headerName: "Имейл",
-      headerAlign: "right",
-      align: "right",
-      flex: 2,
-      minWidth: 300
-    },
-    {
-      field: "group",
-      headerName: "Група",
-      headerAlign: "right",
-      align: "right",
-      flex: 2,
-      minWidth: 300
-    }
-  ];
-
   return {
     rows,
-    columns,
     getLogsLoading,
     setRows
   };
