+++ src/app/doctoralCenter/admin/_lib/UserManagementGridData.jsx
@@ -1,10 +1,9 @@
 import { useEffect, useState } from "react";
-import MenuIcon from "@mui/icons-material/Menu";
-import { Menu, MenuItem } from "@mui/material";
 import DoctoralCenterAdminAPI from "@/api/doctoralCenterAdmin";
 
 import { useSelector } from "react-redux";
 import { selectDoctoralCenter } from "@/features/user/slices/userMemoSelector";
+import { UserManagementColunms } from "../_constants/userManagementColumns";
 
 export default function UserManagementGridData() {
   const [rows, setRows] = useState([]);
@@ -56,68 +55,24 @@ export default function UserManagementGridData() {
     }
   };
 
-  const columns = [
-    { field: "oid", headerName: "Oid", flex: 1.5, minWidth: 200 },
-    {
-      field: "name",
-      headerName: "Име",
-      flex: 1,
-      minWidth: 200
-    },
-    {
-      field: "email",
-      headerName: "Имейл",
-      headerAlign: "right",
-      align: "right",
-      flex: 1,
-      minWidth: 300
-    },
-    {
-      field: "role",
-      headerName: "Роля",
-      headerAlign: "right",
-      align: "right",
-      flex: 2,
-      minWidth: 150
-    },
-    {
-      field: "actions",
-      headerName: "Действия",
-      sortable: false,
-      headerAlign: "center",
-      align: "center",
-      filterable: false,
-      width: 100,
-      renderCell: (params) => {
-        return (
-          <>
-            <MenuIcon
-              onClick={(event) => handleOpenMenu(event, params.row)}
-            ></MenuIcon>
+  const menuItemDisabled = () => {
+    return selectedUser?.oid == doctoralCenter.oid;
+  };
 
-            <Menu
-              anchorEl={menuAnchor}
-              open={menuAnchor}
-              onClose={() => setMenuAnchor(false)}
-            >
-              <MenuItem
-                onClick={() => onMenuClick("delete")}
-                disabled={selectedUser?.oid == doctoralCenter.oid}
-              >
-                Премахни
-              </MenuItem>
-            </Menu>
-          </>
-        );
-      }
-    }
-  ];
+  const { columns } = UserManagementColunms(
+    menuAnchor,
+    setMenuAnchor,
+    handleOpenMenu,
+    onMenuClick,
+    menuItemDisabled
+  );
 
   return {
     rows,
     columns,
     openDialogBoxYesNo,
     setOpenDialogBoxYesNo,
+    columns,
     selectedUser,
     setRows,
     dialogTitle,
